elearning-platform/services/backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
channels>=4.1
channels-redis>=4.2
python-socketio>=5.11
django-rq>=2.10
psycopg[binary]>=3.2
redis>=5.0
boto3>=1.35
cryptography>=43.0
reportlab>=4.2
Pillow>=10.4
uvicorn[standard]>=0.32
numpy>=2.1
opencv-python-headless>=4.10
